# Optimize methods in ground_constant_solver, formatting.